backend/requirements.txt
# Core
Django>=5.0,<6.0
psycopg2-binary>=2.9

# Django REST
djangorestframework>=3.15
django-filter>=24.2

# Auth / JWT
djangorestframework-simplejwt>=5.3

# CORS
django-cors-headers>=4.3

# API Docs
drf-spectacular>=0.27

# Environment variables
python-dotenv>=1.0

# WSGI
gunicorn>=22.0
# Static files
whitenoise>=6.5.0

# Testing
pytest>=7.4
pytest-django>=4.5

ipython>=8.20
django-debug-toolbar>=4.3
cryptography>=42.0